chore(lifed): clippy clean-up — fanout type alias + manual_async_fn + unnecessary_cast (BRO-930)
Three clippy lints surfaced when running clippy --workspace -- -D warnings:

* type_complexity: Vec<mpsc::Sender<Result<pb::AgentEvent, tonic::Status>>>
  in FanoutRegistry — extracted to a pub type AgentEventSender alias.
* manual_async_fn: shutdown_signal returned impl Future; converted to
  async fn now that the body is trivial.
* unnecessary_cast: as_nanos() already returns u128 — drop the cast.

No behaviour change. Unblocks the verify-deps-lifed and Lint CI lanes.

Sub-phase A task A14 prep.

Refs BRO-930

+++ crates/life-runtime/lifed/src/services/agent.rs
@@ -109,7 +109,7 @@ mod uuid_like {
         let nanos = SystemTime::now()
             .duration_since(UNIX_EPOCH)
             .unwrap_or_default()
-            .as_nanos() as u128;
+            .as_nanos();
         nanos.to_le_bytes()
     }
 }
